Update 3.1-7
The abstraction of directories and the handling of files and binaries was refactored simply because the installation script was tested on a UNIX system (XNU/Darwin) and showed inconsistencies in the verification, so it was rewritten in a more aggressive way to identify the location of the dependency binaries.

+++ install.sh
@@ -1,23 +1,23 @@
 #!/bin/bash
 
 hostDependencyChecker(){
-    [[ -f /usr/bin/wine || ! -f /opt/wine-staging/bin/wine ]] && echo -e "$ERROR_LOG Wine not found" && exit
-    [ ! -f /usr/bin/winetricks ] && echo -e "$ERROR_LOG Winetricks not found" && exit
-    [ ! -f /usr/bin/bash ] && echo -e "$ERROR_LOG Bash not found" && exit
-    [ ! -f /usr/bin/wget ] && echo -e "$ERROR_LOG Wget not found" && exit
-    [ ! -f /usr/bin/unzip ] && echo -e "$ERROR_LOG Unzip not found" && exit
-    [ ! -f /usr/bin/tar ] && echo -e "$ERROR_LOG Tar not found" && exit
-    [ ! -f /usr/bin/zenity ] && echo -e "$ERROR_LOG Zenity not found" && exit
+    dependencies=(wine winetricks bash wget unzip tar zenity)
+    for cmd in "${dependencies[@]}"; do
+        if ! command -v "$cmd" >/dev/null 2>&1; then
+            echo -e "$ERROR_LOG $cmd not found" && exit 1
+        fi
+    done
 }
 
 atomicTree(){
     TMP=${TREE}/TMP
     PROGRAM=${TREE}/PROGRAM
     PREFIX=${TREE}/PREFIX
-    [ -d "$PROGRAM" ] && rm -rf $PROGRAM
-    [ -d "$PREFIX" ] && rm -rf $PREFIX
-    [ -d "$TMP" ] && rm -rf $TMP
-    rm -rf $TREE
+    [ -f "$TREE"/TeknoParrot ] && rm -rf "$TREE"/TeknoParrot
+    [ -d "$PROGRAM" ] && rm -rf "$PROGRAM"
+    [ -d "$PREFIX" ] && rm -rf "$PREFIX"
+    [ -d "$TMP" ] && rm -rf "$TMP"
+    mv "$TREE" "$TREE".old
     mkdir -p "$TREE"/{TMP,PROGRAM}
 }
 
@@ -25,31 +25,32 @@ dependencyInstall(){
     clear
     export WINEPREFIX=${PREFIX}
     echo -e "$WAIT_LOG Wineboot." && wineboot -u &> $ARL_LOG
-        [ -d "$PREFIX"/drive_c ] && echo -e " $DONE_LOG Structure created!"
-        [ ! -d "$PREFIX"/drive_c ] && echo -e " $ERROR_LOG Structure not created." && exit
-        echo -e "$WAIT_LOG Downloading dependencies [DXVK & .NET-Runtime/Desktop-Runtime]"
-        wget -c https://aka.ms/dotnet/8.0/dotnet-runtime-win-x64.exe --directory-prefix="$TMP" &>> $ARL_LOG
-        wget -c https://aka.ms/dotnet/8.0/windowsdesktop-runtime-win-x64.exe --directory-prefix="$TMP" &>> $ARL_LOG
-        [ ! -f "$TMP"/dotnet-runtime-win-x64.exe ] && echo -e " $ERROR_LOG .NET Runtime was not downloaded." && exit
-        [ ! -f "$TMP"/windowsdesktop-runtime-win-x64.exe ] && echo -e " $ERROR_LOG .NET Desktop Runtime was not downloaded." && exit
-        winetricks dxvk &>> $ARL_LOG && echo -e " $DONE_LOG DXVK installed!"
-        wine "$TMP"/dotnet-runtime-win-x64.exe /install /quiet /norestart &>> $ARL_LOG && echo -e " $DONE_LOG .NET Runtime installed!"
-        wine "$TMP"/windowsdesktop-runtime-win-x64.exe /install /quiet /norestart &>> $ARL_LOG && echo -e " $DONE_LOG .NET Desktop Runtime installed!"  
-        echo -e "$WAIT_LOG Downloading TeknoParrot (Web-Installer)"
-        wget -c https://github.com/nzgamer41/TPBootstrapper/releases/latest/download/TPBootstrapper.zip --directory-prefix="$TMP" &>> $ARL_LOG
-        [ ! -f "$TMP"/TPBootstrapper.zip ] && echo -e " $ERROR_LOG TPBootstrapper was not downloaded." && exit
-        unzip "$TMP"/TPBootstrapper.zip -d "$PROGRAM" &>> $ARL_LOG
-        [ ! -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $ERROR_LOG TPBootstrapper was not extracted." && exit
-        [ ! -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $ERROR_LOG TPBootstrapper was not found." && exit
-        (
+        if [ -d "$PREFIX"/drive_c ]; then
+            echo -e " $DONE_LOG Structure created!"
+            echo -e "$WAIT_LOG Downloading dependencies."
+            wget -c https://aka.ms/dotnet/8.0/dotnet-runtime-win-x64.exe --directory-prefix="$TMP" &>> $ARL_LOG
+            wget -c https://aka.ms/dotnet/8.0/windowsdesktop-runtime-win-x64.exe --directory-prefix="$TMP" &>> $ARL_LOG
+            winetricks dxvk &>> $ARL_LOG
+            echo -e "$WAIT_LOG Installing dependencies."
+            wine "$TMP"/dotnet-runtime-win-x64.exe /install /quiet /norestart &>> $ARL_LOG
+            wine "$TMP"/windowsdesktop-runtime-win-x64.exe /install /quiet /norestart &>> $ARL_LOG
+            echo -e "$WAIT_LOG Downloading TeknoParrot (Web-Installer)."
+            wget -c https://github.com/nzgamer41/TPBootstrapper/releases/latest/download/TPBootstrapper.zip --directory-prefix="$TMP" &>> $ARL_LOG
+            [ ! -f "$TMP"/TPBootstrapper.zip ] && echo -e " $ERROR_LOG TPBootstrapper was not downloaded." && exit
+            unzip "$TMP"/TPBootstrapper.zip -d "$PROGRAM" &>> $ARL_LOG
+            [ ! -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $ERROR_LOG TPBootstrapper was not extracted." && echo -e " $ERROR_LOG TPBootstrapper was not found." && exit
+            (
             cd "$PROGRAM"
             wine TPBootstrapper.exe &>> $ARL_LOG
             [ -f "$PROGRAM"/TeknoParrotUi.exe ] && echo -e " $DONE_LOG TeknoParrot installed!"
             [ ! -f "$PROGRAM"/TeknoParrotUi.exe ] && echo -e " $ERROR_LOG TeknoParrot not installed." && exit
-        )
-    ls $TREE &>> $ARL_LOG && ls $TREE/* &>> $ARL_LOG
-    rm -rf "$PROGRAM"/TPBootstrapper*
-    rm -rf "$TMP" && echo -e " $DONE_LOG Temporary files cleared!"
+            )
+            ls $TREE &>> $ARL_LOG && ls $TREE/* &>> $ARL_LOG
+            rm -rf "$PROGRAM"/TPBootstrapper*
+            rm -rf "$TMP" && echo -e " $DONE_LOG Temporary files cleared!"
+       else
+            echo -e " $ERROR_LOG Structure not created." && exit
+       fi
 }
 
 executableCreation(){
@@ -57,18 +58,18 @@ executableCreation(){
         cd "$TREE"
         HEADER="#!/bin/bash"
         FLAGS="LC_ALL=C LC_NUMERIC=C LANG=en_US.UTF-8 WINEPREFIX=$PREFIX wine $PROGRAM/TeknoParrotUi.exe"
-        echo $HEADER > TeknoParrot-Linux
-        echo $FLAGS >> TeknoParrot-Linux
-        chmod +x TeknoParrot-Linux
+        echo $HEADER > TeknoParrot
+        echo $FLAGS >> TeknoParrot
+        chmod +x TeknoParrot
     )
 }
 
-ARL_NAME="Arcade Wrapper Linux"
-ARL_VERSION="3.1-6"
+ARL_NAME="Arcade Wrapper Linux/Unix-like"
+ARL_VERSION="3.1-7"
 ARL_LOG="/dev/null"
-DONE_LOG="\e[1;32m*\033[0m"
-WAIT_LOG="\e[1;33m*\033[0m"
-ERROR_LOG="\e[1;31m*\033[0m"
+DONE_LOG="\e[1;32m==>\033[0m"
+WAIT_LOG="\e[1;33m==>\033[0m"
+ERROR_LOG="\e[1;31m==>\033[0m"
 TREE=${HOME}/TeknoParrot
 
 case $1 in
